+++ Sources/Crypto/LEAP.m
@@ -33,12 +33,12 @@
 #include "md4.h"
 
 //calulate the last two bytes
-int gethashlast2(const unsigned char *challenge, const unsigned char *response, unsigned char* endofhash)
+NSInteger gethashlast2(const unsigned char *challenge, const unsigned char *response, unsigned char* endofhash)
 {
     unsigned char zpwhash[7] = { 0, 0, 0, 0, 0, 0, 0 };
     unsigned char cipher[8] = { 0, 0, 0, 0, 0, 0, 0, 0 };
 
-    for (int i = 0; i <= 0xffff; ++i)
+    for (NSInteger i = 0; i <= 0xffff; ++i)
 	{
         zpwhash[0] = i >> 8;
         zpwhash[1] = i & 0xff;
@@ -60,7 +60,7 @@ int gethashlast2(const unsigned char *challenge, const unsigned char *response, 
 
 /* quick wrapper for easy md4 */
 
-void NtPasswordHash(char *secret, int secret_len, unsigned char *hash)
+void NtPasswordHash(char *secret, NSInteger secret_len, unsigned char *hash)
 {
     unsigned char unicodePassword[64];
 
@@ -68,14 +68,14 @@ void NtPasswordHash(char *secret, int secret_len, unsigned char *hash)
     /* This implicitly supports 8-bit ISO8859/1 characters. */
     memset(unicodePassword, 0, sizeof(unicodePassword));
 
-    for (int i = 0; i < secret_len; ++i)
+    for (NSInteger i = 0; i < secret_len; ++i)
         unicodePassword[i * 2] = (unsigned char) secret[i];
 
     /* Unicode is 2 bytes per char */
     md4(unicodePassword, secret_len * 2, hash);
 }
 
-int testChallenge(const unsigned char* challenge, const unsigned char* response, unsigned char *zpwhash) 
+NSInteger testChallenge(const unsigned char* challenge, const unsigned char* response, unsigned char *zpwhash) 
 {
     unsigned char cipher[8] = { 0, 0, 0, 0, 0, 0, 0, 0 };
